Merged PR 719: AISimulator: Add Feature - Filter tensorflow warning in pytest.
The warning in the pytest warning summay is due to the tensorflow. To be precise,
```
/usr/local/lib/python3.5/dist-packages/tensorflow_core/python/pywrap_tensorflow_internal.py:15: PendingDeprecationWarning: the imp module is deprecated in favour of importlib; see the module's documentation for alternative uses
    import imp
```

And it seems that this DeprecationWarning has not been fixed by tensorflow developers yet.
The issue is still open in tf 2.0.
https://github.com/tensorflow/tensorflow/issues/31412
So we can just filter the third-party warning by using pytest.ini configure file.