Merged PR 926: Plan Generator: Add feature - introduces the profiling time into parsed node
**Overview**
This PR mainly introduces the parsed result for profiling database into parsed node, for the convince of AI_simulator. For the file structure, we move the profiler folder into the subfolder of adapter so as our adapter somehow depend on the profiling result to work.
For AI_simulator, future PRs are on their way for nodelist and plan_generator, in order to pass exection_time attr to AI_simulator.

**Main Changes**
1, `tf_parser` :introduce profiling result into parser node
2, folder structure, move the profiler folder into the subfolder of adapter
3, revise the test bench.