Merged PR 1219: SuperScaler: Fix Pipeline - Use single CMD
## Overview

The second CMD will flush the return code the Docker, so if the first CMD return error but the second CMD not, the Docker will not return with error.

But if we use one single CMD and use && to connect two command, it will send error if the first command return will error.